@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "jupyter-to-medium"
-version = "0.2.10"
+version = "0.2.11"
 description = "Publish a Jupyter Notebook as a Medium blogpost"
 authors = ["dexplo <petrou.theodore@gmail.com>"]
 maintainers = ["mjam03 <markjamison03@gmail.com>"]
@@ -10,7 +10,6 @@ homepage = "https://github.com/dexplo/jupyter_to_medium"
 repository = "https://github.com/dexplo/jupyter_to_medium"
 documentation = "https://www.dexplo.org/jupyter_to_medium/"
 classifiers = [
-    "Programming Language :: Python :: 3.7",
     "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "License :: OSI Approved :: MIT License",
@@ -22,13 +21,13 @@ classifiers = [
 Changelog = "https://github.com/dexplo/jupyter_to_medium/releases"
 
 [tool.poetry.dependencies]
-python = ">=3.7,<3.11"
-beautifulsoup4 = "^4.10.0"
-matplotlib = "^3.5.0"
-nbconvert = "5.6.1"
-numpy = "^1.21.4"
-requests = "^2.26.0"
-jupyter-contrib-nbextensions = "^0.5.1"
+python = ">=3.8,<3.11"
+beautifulsoup4 = "^4.12.0"
+matplotlib = "^3.7.1"
+nbconvert = "^7.2.10"
+numpy = "^1.24.2"
+requests = "^2.28.2"
+jupyter-contrib-nbextensions = "^0.7.0"
 jupyter = "^1.0.0"
 
 [tool.poetry.dev-dependencies]
@@ -41,6 +40,10 @@ nox = "^2021.10.1"
 nox-poetry = "^0.9.0"
 
 # let pytest coverage know where code is
+
+[tool.poetry.group.dev.dependencies]
+mypy = "^1.1.1"
+
 [tool.coverage.paths]
 source = ["src", "*/site-packages"]
 
